remove maker accounts instead of reload for wallet disconnect
