+++ src/hooks/useMaker.js
@@ -1,7 +1,6 @@
 import { useContext } from 'react';
 
 import { MakerObjectContext } from '../providers/MakerProvider';
-import { AccountTypes } from '../utils/constants';
 
 function useMaker() {
   const {
@@ -18,37 +17,10 @@ function useMaker() {
     network,
     txLastUpdate,
     connectBrowserProvider,
+    connectToProviderOfType,
+    disconnect,
     navigation
   } = useContext(MakerObjectContext) || {};
-
-  const disconnectWalletLink = subprovider => subprovider.close();
-  const disconnectWalletConnect = async subprovider =>
-    (await subprovider.getWalletConnector()).killSession();
-  const disconnectBrowserWallet = () =>
-    ['lastConnectedWalletType', 'lastConnectedWalletAddress'].forEach(x =>
-      sessionStorage.removeItem(x)
-    );
-
-  const disconnect = () => {
-    const subprovider = maker.service('accounts').currentWallet();
-    if (subprovider.isWalletLink) return disconnectWalletLink(subprovider);
-    else if (subprovider.isWalletConnect) disconnectWalletConnect(subprovider);
-    else if (
-      sessionStorage.getItem('lastConnectedWalletType') ===
-      AccountTypes.METAMASK
-    )
-      disconnectBrowserWallet();
-    document.location.reload();
-  };
-
-  const connectToProviderOfType = async type => {
-    const account = await maker.addAccount({
-      type
-    });
-    maker.useAccountWithAddress(account.address);
-    const connectedAddress = maker.currentAddress();
-    return connectedAddress;
-  };
 
   return {
     maker,
